Reorganize dataset repository extension methods
