+++ kamu-adapter-graphql/src/queries/datasets.rs
@@ -184,7 +184,10 @@ impl Datasets {
     ) -> Result<CreateDatasetFromSnapshotResult> {
         let local_repo = from_catalog::<dyn domain::DatasetRepository>(ctx).unwrap();
 
-        let result = match domain::create_dataset_from_snapshot(local_repo.as_ref(), snapshot).await
+        let result = match local_repo
+            .as_ref()
+            .create_dataset_from_snapshot(snapshot)
+            .await
         {
             Ok(result) => {
                 let dataset = Dataset::from_ref(ctx, &result.dataset_handle.as_local_ref()).await?;
