+++ kamu-adapter-graphql/tests/tests/test_gql_datasets.rs
@@ -64,16 +64,17 @@ async fn dataset_by_id() {
         .build();
 
     let local_repo = cat.get_one::<dyn DatasetRepository>().unwrap();
-    let create_result = create_dataset_from_snapshot(
-        local_repo.as_ref(),
-        MetadataFactory::dataset_snapshot()
-            .name("foo")
-            .kind(DatasetKind::Root)
-            .push_event(MetadataFactory::set_polling_source().build())
-            .build(),
-    )
-    .await
-    .unwrap();
+    let create_result = local_repo
+        .as_ref()
+        .create_dataset_from_snapshot(
+            MetadataFactory::dataset_snapshot()
+                .name("foo")
+                .kind(DatasetKind::Root)
+                .push_event(MetadataFactory::set_polling_source().build())
+                .build(),
+        )
+        .await
+        .unwrap();
 
     let schema = kamu_adapter_graphql::schema(cat);
     let res = schema
