+++ kamu-core/tests/tests/engine/test_ingest_engine.rs
@@ -212,7 +212,9 @@ impl IngestTestHarness {
     }
 
     async fn ingest_snapshot(&self, dataset_snapshot: DatasetSnapshot, dataset_name: &DatasetName) {
-        create_dataset_from_snapshot(self.local_repo.as_ref(), dataset_snapshot)
+        self.local_repo
+            .as_ref()
+            .create_dataset_from_snapshot(dataset_snapshot)
             .await
             .unwrap();
