+++ kamu-core/tests/tests/engine/test_transform_engine.rs
@@ -124,7 +124,9 @@ async fn test_transform_with_engine_spark() {
 
     let root_name = root_snapshot.name.clone();
 
-    create_dataset_from_snapshot(local_repo.as_ref(), root_snapshot)
+    local_repo
+        .as_ref()
+        .create_dataset_from_snapshot(root_snapshot)
         .await
         .unwrap();
 
@@ -156,7 +158,9 @@ async fn test_transform_with_engine_spark() {
 
     let deriv_name = deriv_snapshot.name.clone();
 
-    create_dataset_from_snapshot(local_repo.as_ref(), deriv_snapshot)
+    local_repo
+        .as_ref()
+        .create_dataset_from_snapshot(deriv_snapshot)
         .await
         .unwrap();
 
@@ -347,7 +351,9 @@ async fn test_transform_with_engine_flink() {
 
     let root_name = root_snapshot.name.clone();
 
-    create_dataset_from_snapshot(local_repo.as_ref(), root_snapshot)
+    local_repo
+        .as_ref()
+        .create_dataset_from_snapshot(root_snapshot)
         .await
         .unwrap();
 
@@ -379,7 +385,9 @@ async fn test_transform_with_engine_flink() {
 
     let deriv_name = deriv_snapshot.name.clone();
 
-    create_dataset_from_snapshot(local_repo.as_ref(), deriv_snapshot)
+    local_repo
+        .as_ref()
+        .create_dataset_from_snapshot(deriv_snapshot)
         .await
         .unwrap();
