+++ kamu-core/tests/tests/test_pull_service_impl.rs
@@ -651,14 +651,16 @@ async fn test_set_watermark() {
 
     let dataset_name = n!("foo");
 
-    create_dataset_from_snapshot(
-        harness.local_repo.as_ref(),
-        MetadataFactory::dataset_snapshot()
-            .name(&dataset_name)
-            .build(),
-    )
-    .await
-    .unwrap();
+    harness
+        .local_repo
+        .as_ref()
+        .create_dataset_from_snapshot(
+            MetadataFactory::dataset_snapshot()
+                .name(&dataset_name)
+                .build(),
+        )
+        .await
+        .unwrap();
 
     let num_blocks = || async {
         let ds = harness
@@ -1016,12 +1018,13 @@ impl SyncService for TestSyncService {
                 .unwrap()
             {
                 None => {
-                    create_dataset_from_snapshot(
-                        self.local_repo.as_ref(),
-                        MetadataFactory::dataset_snapshot().name(local_name).build(),
-                    )
-                    .await
-                    .unwrap();
+                    self.local_repo
+                        .as_ref()
+                        .create_dataset_from_snapshot(
+                            MetadataFactory::dataset_snapshot().name(local_name).build(),
+                        )
+                        .await
+                        .unwrap();
                 }
                 Some(_) => (),
             }
